+++ sdk/digitaltwins/Azure.DigitalTwins.Core/src/Models/DigitalTwinComponentMetadata.cs
@@ -8,28 +8,22 @@ namespace Azure.DigitalTwins.Core
 {
     /// <summary>
     /// An optional, helper class for deserializing a digital twin.
-    /// The $metadata class on a <see cref="BasicDigitalTwin"/>.
+    /// The $metadata class on a <see cref="BasicDigitalTwinComponent"/>.
     /// </summary>
     /// <remarks>
     /// For more samples, see <see href="https://github.com/Azure/azure-sdk-for-net/tree/main/sdk/digitaltwins/Azure.DigitalTwins.Core/samples">our repo samples</see>.
     /// </remarks>
-    [JsonConverter(typeof(DigitalTwinMetadataJsonConverter))]
-    public class DigitalTwinMetadata
+    [JsonConverter(typeof(DigitalTwinComponentMetadataJsonConverter))]
+    public class DigitalTwinComponentMetadata
     {
         /// <summary>
-        /// The Id of the model that the digital twin is modeled by.
-        /// </summary>
-        [JsonPropertyName(DigitalTwinsJsonPropertyNames.MetadataModel)]
-        public string ModelId { get; set; }
-
-        /// <summary>
-        /// The date and time the digital twin was last updated.
+        /// The date and time the component was last updated.
         /// </summary>
         [JsonPropertyName(DigitalTwinsJsonPropertyNames.MetadataLastUpdateTime)]
         public string LastUpdatedOn { get; set; }
 
         /// <summary>
-        /// This field will contain metadata about changes on properties on the digital twin.
+        /// This field will contain metadata about changes on properties on the component.
         /// The key will be the property name, and the value is the metadata.
         /// </summary>
 #pragma warning disable CA2227 // Collection properties should be readonly